Refactor app.js and package.json to add axios dependency
